Update README.md
This is an open-source, AI-controlled, solar-powered food synthesis system designed to produce nutritious food from oyster mushrooms in a closed-loop environment. It was conceived not in a lab, but through a necessity born of survival and a refusal to accept a broken system.
my idea is NOT for profit, but to help end world hunger, in theory it removes all need for farming, agriculture and killing of any animals. I beleive my idea is a world changer, but as a poor homeless man living in a country run on profits alone, I'm very limited to what I'm able to actually do, I have had this detailed plan for about 3 years so far unsuccessfully trying to get this idea seen and recognised